Incorporate Tailwind v3
CAUTION: in January 2025 Tailwind v4 has been released, which uses a
different installation and config method than Tailwind v3 (see [1]). To
keep using Tailwind v3, it must be installed and configured as follows:

    npm install -D tailwindcss@3 postcss autoprefixer
    npx tailwindcss init -p

When omitting the '@3' modifier, Tailwind v4 will be installed which
causes the 'npx tailwindcss init -p' to fail with a 'could not determine
executable to run' error because Tailwind v4 doesn't define an 'init'
command.

[1] https://tailwindcss.com/docs/upgrade-guide