Include GitHub API token in githubUserInfo store
Instead of treating it as a separate persistent value in localStorage.
This could be implemented with very little effort since the
githubUserInfo store is already persisted in localStorage, and the
GitHub user info and the API token are obtained at the same time and
have the same lifecycle.

This also makes githubApiTokenStore.ts and any dedicated functions for
handling the GitHub API token obsolete.